Moved image display component up one directory.
