+++ PhotoPlusFrontend/PhotoPlus/src/app/components/image-display/image-display.component.html
